refactor(canvas): 拆出拖动松手写回边界
